+++ src/Senparc.Xncf.ExtensionAreaTemplate/Areas/Admin/Pages/MyApp/About.cshtml.cs
@@ -1,12 +1,5 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
-using Senparc.Ncf.Core.Models.DataBaseModel;
 using Senparc.Ncf.Service;
-using Senparc.Ncf.XncfBase;
+using System;
 
 namespace Senparc.Xncf.ExtensionAreaTemplate.Areas.MyApp.Pages
 {
