+++ src/Senparc.Xncf.ExtensionAreaTemplate/Areas/Admin/Pages/_ViewImports.cshtml
