+++ src/Senparc.Xncf.ExtensionAreaTemplate/Areas/Admin/Pages/_ViewStart.cshtml
